Manual merge of typecheckcleanup200810 branch as of rev 731346 in that branch, resolved all conflicts; with this BigDecimal is not the native type for many entity fields and is used for many calculations; some work is still needed in certain areas of the project but this covers the most critical ones; thanks especially to Scott for his work on this
git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@731851 13f79535-47bb-0310-9956-ffa450edef68

+++ applications/product/src/org/ofbiz/product/spreadsheetimport/ImportProductHelper.java
@@ -19,6 +19,7 @@
 
 package org.ofbiz.product.spreadsheetimport;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 import javolution.util.FastMap;
@@ -46,15 +47,15 @@ public class ImportProductHelper {
 
     // prepare the inventoryItem map
     public static Map<String, Object> prepareInventoryItem(String productId,
-            double quantityOnHand, String inventoryItemId) {
+            BigDecimal quantityOnHand, String inventoryItemId) {
         Map<String, Object> fields = FastMap.newInstance();
         fields.put("inventoryItemId", inventoryItemId);
         fields.put("inventoryItemTypeId", "NON_SERIAL_INV_ITEM");
         fields.put("productId", productId);
         fields.put("ownerPartyId", "Company");
         fields.put("facilityId", "WebStoreWarehouse");
-        fields.put("quantityOnHandTotal", Double.valueOf(quantityOnHand));
-        fields.put("availableToPromiseTotal", Double.valueOf(quantityOnHand));
+        fields.put("quantityOnHandTotal", quantityOnHand);
+        fields.put("availableToPromiseTotal", quantityOnHand);
         return fields;
     }
